requirements.txt
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.21
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0

# Embedding 推理 (替代 torch + sentence-transformers)
onnxruntime>=1.16.0
tokenizers>=0.13.0

httpx>=0.27.0
trafilatura>=1.6.0

mss>=9.0.0
pygetwindow>=0.0.9
